requirements.txt
flask
psycopg2-binary
flask-sqlalchemy
bcrypt
python-dotenv
gunicorn
opencv-python-headless
